ISSUE-612 refactored triggers to be more robust, tied to repository methods instead of explicit service calls
